\name{write_mol2}
\alias{write_mol2}
\title{Saves Sybyl mol2-file}
\description{
  Writes file in Sybyl mol2 format.
}
\usage{
write_mol2(filename)
}
\arguments{
	\item{filename}{the name of the mol2 file to save}
}
\value{Molecular database (mdb) list}
\examples{
## Save mol2 file
\dontrun{
write_mol2("fname.mol2")
}
}
